rworker handles variable messages with a single variableresponse class with unit tests

+++ src/main/java/edu/wvu/stat/rc2/rworker/message/VariableUpdateMessage.java
@@ -7,13 +7,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class VariableUpdateMessage extends BaseMessage {
 	private final Map<String,Object> _variables;
-	private final long _userIdentifier;
+	private final int _userIdentifier;
 	private final boolean _isDelta;
 	
 	@JsonCreator
 	public VariableUpdateMessage(
 			@JsonProperty("msg") String msg,
-			@JsonProperty("userIdentifier") long userIdent,
+			@JsonProperty("userIdentifier") int userIdent,
 			@JsonProperty("variables") Map<String,Object> vars,
 			@JsonProperty("delta") boolean delta
 			) 
@@ -24,7 +24,7 @@ public class VariableUpdateMessage extends BaseMessage {
 		_userIdentifier = userIdent;
 	}
 	
-	public long getUserIdentifier() { return _userIdentifier; }
+	public int getUserIdentifier() { return _userIdentifier; }
 	public Map<String,Object> getVariables() { return _variables; }
 	public boolean isDelta() { return _isDelta; }
 }
